Added Functions and Invoke for Diag
Added functions to convert chat to int and print to the diag text box.
the print function is mainly to be invoked by the background worker.
The LineFollow button now communicate with Sparki and reports back data
from left right and center line sensors and reports these values to the
DIAG text box.  Consider creating a header file to clean up code.